requirements.txt
pandas
PyYAML
openai
IPython
deepmerge
spacy
nltk
matplotlib
beautifulsoup4
lxml
en_core_web_sm @  https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl